% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_ms}
\alias{format_ms}
\title{Format summaries for tables}
\usage{
format_ms(x, figs = 3, sigfig = TRUE, use_quantiles = FALSE)
}
\arguments{
\item{x}{Numeric vector to be summarised}

\item{figs}{Number of digits or significant figure to print}

\item{sigfig}{If \code{TRUE}, \code{figs} is interpreted as significant figures;
if \code{FALSE}, \code{figs} is interpreted as number of digits. If significant
figures are used, these are applied to the mean or median. Standard deviation or
first/third quartiles are printed with the same number of digits as the mean
or median.}

\item{use_quantiles}{If \code{TRUE}, output is median (IQR). If \code{FALSE},
output is mean (SD).}
}
\value{
A string, formatted as median (IQR) or mean (SD) with the chosen
  number of digits / significant figures, suitable for display in tables.
}
\description{
Format summaries for tables
}
