% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{whozr}
\alias{whozr}
\title{Calculate Z-scores from LMS tables}
\usage{
whozr(y, x, sex, ref, adjust_large_z = FALSE)
}
\arguments{
\item{y}{Outcome veariable}

\item{x}{Predictor, typically age}

\item{sex}{Sex, coded as "F" and "M"}

\item{ref}{Reference data. A tibble or data.frame with columns named \code{sex},
\code{x},\code{l}, \code{m} and \code{s}}
}
\value{
A vector of \code{y}-for-\code{x} z-scores
}
\description{
Calculate z-scores from LMS tables. Wrappers are provided for use with the WHO
  growth standards (0-5 years) and WHO growth references (5-19 years): see
  \code{\link{zwa}}, \code{\link{zha}}, \code{\link{zwh}}, \code{\link{zwl}},
  \code{\link{zba}}, \code{\link{zhca}}, \code{\link{zaca}}, \code{\link{ztsa}},
  \code{\link{zssa}}
}
