% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zba}
\alias{zba}
\title{BMI-for-age z-scores}
\usage{
zba(bmi, age, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{bmi}{BMI}

\item{age}{Age (days)}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZBA > 5 or ZBA < -5), described
as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of BMI-for-age z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years) and WHO
  growth references (5-19 years).
}
