% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zha}
\alias{zha}
\title{Height-for-age z-scores}
\usage{
zha(height, age, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{height}{Height (cm), measured recumbent up to 2 years (day 730) and
standing thereafter, as recommended by WHO.}

\item{age}{Age (days)}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZHA > 6 or ZHA < -6), described
as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of height-for-age z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years) and WHO
  growth references (5-19 years). Like its alias \code{\link{zla}}:
  both assume that length was measured recumbent up to 2 years (day 730) and
  height was measured standing from day 731 onward.
}
