% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zhca}
\alias{zhca}
\title{Head circumference-for-age z-scores}
\usage{
zhca(headc, age, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{headc}{Head circumference (cm)}

\item{age}{Age (days)}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZHCA > 5 or ZHCA < -5), described
as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of head circumference-for-length z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years).
}
