% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zma}
\alias{zma}
\title{MUAC-for-age z-scores}
\usage{
zma(muac, age, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{muac}{Mid-upper arm cicrumference (cm)}

\item{age}{Age (days)}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZMA > 5 or ZMA < -5), described
as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of MUAC-for-length z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years).
}
