% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zwa}
\alias{zwa}
\title{Weight-for-age z-scores}
\usage{
zwa(weight, age, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{weight}{Weight (kg)}

\item{age}{Age (days)}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZWA > 5 or ZWA < -6),
described as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of weight-for-age z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years) and WHO
  growth references (5-10 years).
}
