% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zwh}
\alias{zwh}
\title{Weight-for-height z-scores}
\usage{
zwh(weight, height, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{weight}{Weight (kg)}

\item{height}{Height (cm), measured standing}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZWH > 5 or ZWH < -5), described
as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of weight-for-height z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years).
}
