% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zwl}
\alias{zwl}
\title{Weight-for-length z-scores}
\usage{
zwl(weight, length, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{weight}{Weight (kg)}

\item{length}{Length (cm), measured recumbent}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZWL > 5 or ZWL < -5), described
as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of weight-for-length z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years).
}
