% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zla}
\alias{zla}
\title{Length-for-age z-scores}
\usage{
zla(length, ...)
}
\arguments{
\item{length}{Length (cm), measured recumbent up to 2 years (day 730) and
standing thereafter, as recommended by WHO.}

\item{...}{Other arguments to \code{\link{zha}}.}
}
\value{
A vector of height-for-age z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years) and WHO
  growth references (5-19 years). This is just an alias for
  \code{\link{zha}}: both assume that length was measured recumbent up to 2
  years (day 730) and height was measured standing from day 731 onward.
}
