% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zssa}
\alias{zssa}
\title{Subscapular skinfold thickness-for-age z-scores}
\usage{
zssa(subscap, age, sex, trim_extreme_z = FALSE)
}
\arguments{
\item{subscap}{Subscapular skinfold thickness (mm)}

\item{age}{Age (days)}

\item{sex}{Sex, coded as "F" and "M"}

\item{trim_extreme_z}{Replace extreme scores (ZSSA > 5 or ZSSA < -5),
described as biologically implausible by WHO, with \code{NA}.}
}
\value{
A vector of subscapular skinfold thickness-for-length z-scores
}
\description{
Calculate z-scores using with the WHO growth standards (0-5 years).
}
